Rename addComments.js to commentApi.js
